autophot/requirements.txt
aafitrans==0.2.1
astroalign==2.4.1
astroplan==0.9.1
astropy==5.3.4
astroquery==0.4.6
astroscrappy==1.2.0
image_registration==0.2.8
lmfit==1.2.2
mastcasjobs==0.0.6
matplotlib==3.9.2
numpy==2.1.0
pandas==2.2.2
photutils==1.11.0
PyYAML==6.0.1
PyYAML==6.0.2
PyZOGY==0.0.1
reproject==0.9
Requests==2.32.3
scipy==1.14.1
sfft==1.5.0
skimage==0.0
